Implement voids (empty ignored spaces)
Voids are empty spaces in the graph. WFC does not propagate constraints to or
from voids, similar to how it ignores walls if the --wrapping options is not
supplied. To designate a slot as a void one needs to place a "?" character in
the initial state TXT.

To simplify the implementation of voids (being able to use 0 as the void flag),
name interning was also implemented. Module names in the CSV and initial state
confifuration no longer need to start with 0 or have no gaps. Instead they can
now be any strings and are interned on import. Export maintains the names from input.