Use u64::checked_shl with overflow handling instead of the ops::Shl (#9)
Panics in debug builds were caused by shifting out all 64 bits out of an
u64. Doing this causes a panic when overflow checks are enabled (they are by
default enabled in debug builds) and produces an unintuitive result in release
builds - it wraps the RHS of the shift operator instead of shifting out all the
bits, so `u64::MAX << 64 == u64::MAX`.

It was a bit tricky to determine whether the fix would change output, but as far
as I understand there should no change.

The faulty logic was only used for iterator size hints. Usually that would mean
no change in output, because they are not to be relied upon for correctness by
iterator implementations. While in the `choose_random()` function we do use
iterator size hints also to apply modulus on the random number we pick, we never
pass anything but a freshly created iterator from our bit vector there so the
unintuitive shifting fallback in production mode shouldn't have manifested
there. I also didn't find any change in output when comparing on sample data in
this repo.