Optimize bitvec and use LTO (#14)
This speeds up the solver by 34% (measured in macro-benchmarks with time `cargo run --release ./samples/sample3.csv -x 100 -y 100 -z 100`) . The speedup is gained from bitvec optimization and tweaking build profiles.

The bitvec optimization was guided by profiling and amounts to a 28% speedup. `BitVec::len` is called very often by the solver, and internally it was calling u64::count_ones intrinsics which showed up significantly on the flamegraph. The patch changes `BitVec::len` to instead return a cached length of the bit vector. The length is stored in the last eight bits of data, so our module count decreased from 256 to 248. A couple of unused methods were removed from `BitVec` so that they didn't have to be adapted to the new data layout.

Tweaking the release build profile amounts to another 6% speedup, but that may not be as prominent for the built DLL as it is for the CLI testing app.