Slot weights (#21)
This adds per-slot, per-module weights.

They can be set with wfc_world_state_slot_module_weights_set. The weights must be normal (not NaN, not infinite, not denormal), positive floats. By itself, this doesn't do anything, but the solver can be initialized with two feature flags, that will cause it to use the weights. These are:

- Features::WEIGHTED_ENTROPY computes Shannon Entropy of the slot using the provided weights. This is quite expansive to do, because all slots need to be considered when observing, and the computation needs to access a lot of memory and perform float maths.

- Features::WEIGHTED_OBSERVATION uses the weights to skew the random choice of a module in a slot during observation, and is quite cheap to do, because it is only computed once per observation.